Implement Windows Hook
Use hook for CapsLock interception instead of the registered hotkey because it worked really buggy, while this works very well.

Still use the registered hotkeys for the other actions where this is sufficient.Implement Windows Hook

Use hook for CapsLock interception instead of the registered hotkey because it worked really buggy, while this works very well.

Still use the registered hotkeys for the other actions where this is sufficient.